Constrain pred ctor param types and casts to 'any'
VeriFast now checks that a predicate constructor parameter type
does not contain a predicate type in a negative position, and that
a type being cast to 'any' does not contain 'any' in a negative
position.

Furthermore, VeriFast allows predicate constructor parameter types to
contain 'any', but only in positive positions, and it allows types that
are cast to 'any' to contain predicate types, but only in positive
positions.